requirements.txt
pandas
numpy
scikit-learn
kafka-python
mysql-connector-python
python-dotenv
joblib
matplotlib
seaborn
miceforest